bugs fix  - video recorded is less than the duration defined  - not set to MONITOR mode when the system is activated due to scheduling

+++ src/include/EagleEye.hpp
@@ -39,7 +39,7 @@ protected:
   Status status;
 
   unsigned int cam_id;
-  unsigned int record_sec;
+  float record_sec;
   myMat frame;
   Camera cam;
 };
